feat(db-core): 카운트 칼럼 타입 수정

+++ core/core-api/src/main/java/com/imstargg/core/api/controller/v1/response/BattleEventBrawlerResultStatisticsResponse.java
@@ -4,10 +4,10 @@ import com.imstargg.core.domain.statistics.BattleEventBrawlerResultStatistics;
 
 public record BattleEventBrawlerResultStatisticsResponse(
         long brawlerBrawlStarsId,
-        int victoryCount,
-        int defeatCount,
-        int drawCount,
-        int starPlayerCount
+        long victoryCount,
+        long defeatCount,
+        long drawCount,
+        long starPlayerCount
 ) {
 
     public static BattleEventBrawlerResultStatisticsResponse of(BattleEventBrawlerResultStatistics statistics) {
